requirements.txt
Flask>=2.0
gunicorn
Flask-SQLAlchemy
psycopg2-binary
python-dateutil
pytz
openai
